+++ database/migrations/2018_06_08_111507_add_in_name_of_to_donation.php
@@ -1,0 +1,32 @@
+<?php
+
+use Illuminate\Support\Facades\Schema;
+use Illuminate\Database\Schema\Blueprint;
+use Illuminate\Database\Migrations\Migration;
+
+class AddInNameOfToDonation extends Migration
+{
+    /**
+     * Run the migrations.
+     *
+     * @return void
+     */
+    public function up()
+    {
+        Schema::table('donations', function (Blueprint $table) {
+            $table->text('in_name_of')->nullable()->after('reference');
+        });
+    }
+
+    /**
+     * Reverse the migrations.
+     *
+     * @return void
+     */
+    public function down()
+    {
+        Schema::table('donations', function (Blueprint $table) {
+            $table->dropColumn('in_name_of');
+        });
+    }
+}