+++ resources/views/fundraising/donations/export-single.blade.php
@@ -5,6 +5,8 @@
             <th>@lang('fundraising.channel')</th>
             <th>@lang('fundraising.purpose')</th>
             <th>@lang('fundraising.reference')</th>
+            <th>@lang('fundraising.in_name_of')</th>
+            <th>@lang('fundraising.thanked')</th>
             <th>@lang('app.amount')</th>
             <th>@lang('fundraising.exchange_amount')</th>
         </tr>
@@ -16,6 +18,8 @@
                 <td>{{ $donation->channel }}</td>
                 <td>{{ $donation->purpose }}</td>
                 <td>{{ $donation->reference }}</td>
+                <td>{{ $donation->in_name_of }}</td>
+                <td>{{ optional($donation->thanked)->toDateString() }}</td>
                 <td>{{ $donation->amount }}</td>
                 <td>{{ $donation->exchange_amount }}</td>
             </tr>